backtest-worker/worker/celery_app.py
import os
from celery import Celery
from dotenv import load_dotenv

load_dotenv()

app = Celery(
    "backtest_worker",
    broker=os.getenv("REDIS_URL", "redis://localhost:6379/0"),
    backend=os.getenv("REDIS_URL", "redis://localhost:6379/0"),
)

app.conf.update(
    task_serializer="json",
    result_serializer="json",
    accept_content=["json"],
    task_track_started=True,
    worker_prefetch_multiplier=1,
    task_time_limit=600,
    task_soft_time_limit=540,
    result_expires=86400,
)

app.autodiscover_tasks(["worker"])
