backtest-worker/requirements.txt
celery==5.4.0
redis==5.0.8
pandas==2.2.2
pandas-ta==0.3.14b
numpy==1.26.4
python-dotenv==1.0.1
docker==7.1.0
